+++ src/components/GlobalStyle.tsx
@@ -3,9 +3,16 @@ import styled from 'styled-components/native';
 import Layout from '../constans/Layout';
 import Colors from '../constans/Colors';
 
-const CustomView = styled.View`
+
+//add flex
+export const CustomView = styled.View`
 	margin-top: ${Layout.statusBar}px;
 	height: ${Layout.window.height}px;
 `;
 
-export default CustomView;
+export const CustomText = styled.Text`
+	display: flex;
+	text-align: center;
+	color: ${Colors.black};
+	font-size: 20px;
+`;