+++ src/screens/Welcome/index.tsx
@@ -1,20 +1,17 @@
 import React, { FC } from 'react';
-import { Text, View, Button } from 'react-native';
+import {Image, Text, View, Button } from 'react-native';
 import styled from 'styled-components/native';
 import { useNavigation } from '@react-navigation/native';
 
 import Colors from '../../constans/Colors';
-import Layout from '../../constans/Layout';
+
+import * as Global from '../../components/GlobalStyle';
+import foto from '../../assets/foto.jpg';
 
 const WelcomeText = styled.Text`
     margin: 120px 20px;
     font-size: 16px;
     color: ${Colors.black};
-`;
-
-//make global view style for TOP margin
-const HomeView = styled.View` 
-	margin-top: ${Layout.statusBar}px;
 `;
 
 interface IWelcomeProps {
@@ -24,14 +21,15 @@ interface IWelcomeProps {
 const Welcome: FC<IWelcomeProps> = props => {
 	const navi = useNavigation();
     return (
-        <HomeView>
-			<Text>Witam witam </Text>
+        <Global.CustomView>
+			<Global.CustomText>Witam witam </Global.CustomText>
+			<Image source={foto} style={{ width: 100, height: 100}}></Image>
             <WelcomeText>Welcome {props.myProps}</WelcomeText>
 			<Button title="Second Page" onPress={() => {
 				navi.navigate('Second');
 			}}
 			/>
-        </HomeView>
+        </Global.CustomView>
     );
 };